git-svn-id: https://www.ekylibre.org/svn/trunk/ekylibre@936 67a09383-3dfa-4221-8551-890bac9c277c

+++ db/migrate/20090720100733_sep1b3.rb
@@ -9,13 +9,12 @@ class Sep1b3 < ActiveRecord::Migration
       table = model_name.pluralize.to_sym
       model = model_name.camelcase.constantize
       if model.table_exists?
-        add_column table, :creator_id, :integer
-        add_column table, :updater_id, :integer
-        if model.columns_hash.keys.include? "created_by"
-          model.update_all('creator_id=created_by, updater_id=updated_by')
-          remove_column table, :created_by
-          remove_column table, :updated_by 
-        end
+        add_column table, :creator_id, :integer unless model.columns_hash.keys.include? "creator_id"
+        add_column table, :updater_id, :integer unless model.columns_hash.keys.include? "updater_id"
+        model.update_all('creator_id=created_by') if model.columns_hash.keys.include?("creator_id") and model.columns_hash.keys.include?("created_by")
+        model.update_all('updater_id=updated_by') if model.columns_hash.keys.include?("updater_id") and model.columns_hash.keys.include?("updated_by")
+        remove_column table, :created_by if model.columns_hash.keys.include? "created_by"
+        remove_column table, :updated_by if model.columns_hash.keys.include? "updated_by"
       end
     end
     add_column :languages, :created_at, :timestamp
@@ -29,13 +28,12 @@ class Sep1b3 < ActiveRecord::Migration
       table = model_name.pluralize.to_sym
       model = model_name.camelcase.constantize
       if model.table_exists?
-        add_column table, :created_by, :integer
-        add_column table, :updated_by, :integer
-        if model.columns_hash.keys.include? "creator_id"
-          model.update_all('created_by=creator_id, updated_by=updater_id')
-          remove_column table, :creator_id
-          remove_column table, :updater_id
-        end
+        add_column table, :created_by, :integer unless model.columns_hash.keys.include? "created_by"
+        add_column table, :updated_by, :integer unless model.columns_hash.keys.include? "updated_by"
+        model.update_all('created_by=creator_id') if model.columns_hash.keys.include?("creator_id") and model.columns_hash.keys.include?("created_by")
+        model.update_all('updated_by=updater_id') if model.columns_hash.keys.include?("updater_id") and model.columns_hash.keys.include?("updated_by")
+        remove_column table, :creator_id if model.columns_hash.keys.include?("creator_id")
+        remove_column table, :updater_id if model.columns_hash.keys.include?("updater_id")
       end
     end
     remove_column :languages, :updated_by
